requirements.txt
python-telegram-bot>=3.1.2
mock>=1.3.0
howdoi>=1.1.7
PyYAML==3.11
ndg-httpsclient>=0.4.0
pyOpenSSL>=0.15.1